model: use unsigned int for node surfaces

+++ include/model.h
@@ -189,8 +189,8 @@ typedef struct mnode_s {
     mplane_t *plane;
     struct mnode_s *children[2];
 
-    unsigned short firstsurface;
-    unsigned short numsurfaces;
+    unsigned int firstsurface;
+    unsigned int numsurfaces;
 } mnode_t;
 
 /* forward decls; can't include render.h/glquake.h */
